@@ -12,16 +12,22 @@
     "@contentful/rich-text-react-renderer": "^15.12.1",
     "contentful": "^9.2.9",
     "framer-motion": "^7.6.6",
-    "next": "12.2.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^13.4.13",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-responsive-masonry": "^2.1.6",
-    "sass": "^1.55.0"
+    "sass": "^1.55.0",
+    "sharp": "^0.32.4"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
+    "@types/node": "20.4.9",
+    "@types/react": "^18.2.20",
+    "autoprefixer": "^10.4.14",
     "eslint": "8.22.0",
-    "eslint-config-next": "12.2.5",
-    "styled-components": "^5.3.5"
+    "eslint-config-next": "^13.4.13",
+    "postcss": "^8.4.27",
+    "tailwindcss": "^3.3.3",
+    "typescript": "5.1.6"
   }
 }
